moved json data to individual files
I decided it's probably best to have each table's data in its own json
file. While it isn't a problem now, it will get out of hand very quickly
if all the data is in one file. This will also allow individual tables to
be reimported instead of tossing the whole database if such a need should
arise.