added the ability to give enemies abilities.
more tweaking will need to be done as abilities will need to be taggable and more things will need to be in the dropdown menus as I go on. I'm not completely happy having all the ability functionality in the enemy package, as it's making that package much less of a package by being multipurpose. the problem is that abilities are only assigned to enemies (players have skills...is my current thought process) and it seems silly to have an entire package dedicated to that. I'm going to think about that a lot, and eventually I might refactor that into a new form.

as more of a note to myself, I'm considering making a Taggable abstract model since I use that so often and it doesn't really need to be unique.