Added enemy importing and fixed up the enemy view a bit.
There was a lot that went in to this. A lot of frustration was had, and some of this might not be great. But it works, and for now that's good enough. In order to finish up the enemy view, I need to spend another day or two in data entry getting all the items imported.