massive step towards the future.
I've spent the past few days learning (and getting very angry at) several Javascript options. I'm familiar with backbone (though I didn't know how to set it up, hence part of my frustration), but I looked in to Angular2 and React as well. I've settled on backbone though, as it doesn't require me to learn a new syntax (JSX or Typescript). I'm porting the creation forms to be views, so that part is not quite working yet. Basically I added a bunch of groundwork, and it will start working soon.