fixed some errors that arose from removing backbone-forms.
I also added a bit more of the enemy creation form. It isn't done yet and
I want to come up with a better style for it, so this is still a WIP.
Note: any time you see the color #FF00FF (magenta) on an element, that's
just me testing something. That will never be an actual color in
production.