added some more enemy data.
I've changed my opinion on adding data here vs waiting for the form to be
done. Although the form will be easier to use, it will still take a while
to implement. Adding all this data via JSON also takes forever, so I'm not
sure which is worse. But I will probably enter some data manually every
now and then at the very least. It's the perfect thing to do when I want
to be productive, but only for a few minutes.